Unique game files added.
If you wish to edit then simply download this repository open the uproject the other folders and junk you might expect to be there is automatically generated by unreal on your first open. Once you have finished submit a pull request to the repo and ill check for any problems with deleted files ect and commit to the master branch where the file will be integrated into the existing code.